% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbifniche_package.R
\docType{package}
\name{gbifniche}
\alias{gbifniche}
\title{gbifniche: R tool for assessing uncertainties around qualities of observation for environmental niches using GBIF data.}
\description{
R tool for assessing uncertainties around qualities of observation for environmental niches using GBIF data.
}
