% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_gbif}
\alias{read_gbif}
\title{function to read gbif with the option to fetch}
\usage{
read_gbif(species = "Carcharodon carcharias", refresh = FALSE, dwc = TRUE)
}
\arguments{
\item{species}{character, latin name of species}

\item{refresh}{logical, if true fetch fresh set of data}
}
\value{
dwc logical, if TRUE trim to the recommended Darwin Core content

data frame in the form of a tibble
}
\description{
reads file by species name, if file doesn't exist, first try to fetch it then save it to the cache
}
