% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rename_gbif}
\alias{rename_gbif}
\title{Rename one or more fields (columns) in a GBIF tibble}
\usage{
rename_gbif(
  x,
  fields = c(name = "scientificName", longitude = "decimalLongitude", latitude =
    "decimalLatitude")
)
}
\arguments{
\item{x}{tibble of data}

\item{fields}{named character vector with oldname = newname}
}
\value{
renamed tibble
}
\description{
Rename one or more fields (columns) in a GBIF tibble
}
