% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_gbif}
\alias{fetch_gbif}
\title{fetch occurrence data for given species from GBIF}
\usage{
fetch_gbif(species = "Carcharodon carcharias", cache = TRUE)
}
\arguments{
\item{species}{character, latin name of species to fetch}

\item{cache}{logical, if true save results to cache}
}
\value{
data frame in the form of a tibble
}
\description{
fetch occurrence data for given species from GBIF
}
